requirements.txt
Django==4.1.5
django-extensions==3.2.1
django-model-utils==4.3.1
djangorestframework==3.14.0
